REF: started component refactoring, added Loading-Spinner library and loading component
