+++ README.md
@@ -47,6 +47,7 @@
 [✔] make sure that you cannot add the same ingredientblueprint to the same meal more than one time
 [✔] Move all types into types.ts
 [✔] turn classes into typed objects instead
+[] maybe move the unitById and ingredientToId to context and global state?
 [] Comportmentalize the app
 [] copy meals
 [] move meals to different day
