+++ package.json
@@ -20,6 +20,7 @@
     "react-calendar": "^5.1.0",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.54.2",
+    "react-loader-spinner": "^6.1.6",
     "react-router": "^7.1.1",
     "uuid": "^11.0.5"
   },
