+++ src/components/FormComponents/SubmitInputElement.tsx
@@ -1,5 +1,9 @@
 export default function SubmitInputElement({ submitInputText = "Submit" }: { submitInputText?: string }) {
     return (
-        <input className=" cursor-pointer bg-slate-200 rounded-md hover:bg-slate-400 active:bg-slate-600" type="submit" value={submitInputText} />
+        <input
+            className=" cursor-pointer bg-slate-200 rounded-md hover:bg-slate-400 active:bg-slate-600"
+            type="submit"
+            value={submitInputText}
+        />
     )
 }
