+++ src/components/Global/Header.tsx
@@ -1,0 +1,6 @@
+
+export default function Header({ text }: { text: string }) {
+    return (
+        <div className="mx-auto w-full text-center text-5xl m-10">{text}</div>
+    )
+}
