+++ src/components/Global/Loading.tsx
@@ -1,0 +1,9 @@
+import { InfinitySpin } from "react-loader-spinner";
+
+export default function Loading() {
+    return (<InfinitySpin
+        width="200"
+        color="aqua"
+    />)
+
+}
